V3/pyproject.toml
[tool.poetry]
name = "kizmeow-opensea-and-etherscan-discord-bot"
version = "0.1.0"
description = ""
authors = ["c0mrade <killmy666@gmail.com>"]
readme = "README.md"
packages = [{include = "kizmeow_opensea_and_etherscan_discord_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
py-cord = "^2.2.2"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
etherscan-python = "^2.1.0"
slug = "^2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
